+++ class/class.veritrans-gateway-binpromo.php
@@ -6,17 +6,17 @@
     /**
        * Veritrans Payment Gateway Class
        */
-    class WC_Gateway_Veritrans_Installment extends WC_Payment_Gateway {
+    class WC_Gateway_Veritrans_Binpromo extends WC_Payment_Gateway {
 
       /**
        * Constructor
        */
       function __construct() {
-        $this->id           = 'veritrans_installment';
+        $this->id           = 'veritrans_binpromo';
         $this->icon         = apply_filters( 'woocommerce_veritrans_icon', '' );
-        $this->method_title = __( 'Veritrans Installment', 'colabsthemes' );
+        $this->method_title = __( 'Veritrans BIN Promo', 'colabsthemes' );
         $this->has_fields   = true;
-        $this->notify_url   = str_replace( 'https:', 'http:', add_query_arg( 'wc-api', 'WC_Gateway_Veritrans_Installment', home_url( '/' ) ) );
+        $this->notify_url   = str_replace( 'https:', 'http:', add_query_arg( 'wc-api', 'WC_Gateway_Veritrans_Binpromo', home_url( '/' ) ) );
 
         // Load the settings
         $this->init_form_fields();
@@ -32,12 +32,13 @@
         $this->environment        = $this->get_option( 'select_veritrans_environment' );
         $this->to_idr_rate        = $this->get_option( 'to_idr_rate' );
         $this->enable_sanitization = $this->get_option( 'enable_sanitization' );
-        $this->enable_bni         = $this->get_option( 'enable_bni_installment' );
-        $this->bni_terms          = $this->get_option( 'enable_bni_installment_terms' );
-        $this->enable_mandiri     = $this->get_option( 'enable_mandiri_installment' );
-        $this->mandiri_terms      = $this->get_option( 'enable_mandiri_installment_terms' );
+        // $this->enable_bni         = $this->get_option( 'enable_bni_installment' );
+        // $this->bni_terms          = $this->get_option( 'enable_bni_installment_terms' );
+        // $this->enable_mandiri     = $this->get_option( 'enable_mandiri_installment' );
+        // $this->mandiri_terms      = $this->get_option( 'enable_mandiri_installment_terms' );
+        // $this->installment_terms      = $this->get_option( 'installment_terms' );
         $this->bin_filter         = $this->get_option( 'enable_bin_filter' );
-        $this->min_amount         = $this->get_option( 'min_amount' );
+        // $this->min_amount         = $this->get_option( 'min_amount' );
 
 
         $this->log = new WC_Logger();
@@ -73,8 +74,8 @@
        * @return void
        */
       public function admin_options() { ?>
-        <h3><?php _e( 'Veritrans Installment', 'woocommerce' ); ?></h3>
-        <p><?php _e('Allows installment payments using Veritrans.', 'woocommerce' ); ?></p>
+        <h3><?php _e( 'Veritrans BIN Promo', 'woocommerce' ); ?></h3>
+        <p><?php _e('Allows BIN promo payments using Veritrans.', 'woocommerce' ); ?></p>
         <table class="form-table">
           <?php
             // Generate the HTML For the settings form.
@@ -97,21 +98,21 @@
           'enabled' => array(
             'title' => __( 'Enable/Disable', 'woocommerce' ),
             'type' => 'checkbox',
-            'label' => __( 'Enable Veritrans installment Payment', 'woocommerce' ),
+            'label' => __( 'Enable Veritrans BIN promo Payment', 'woocommerce' ),
             'default' => 'no'
           ),
           'title' => array(
             'title' => __( 'Title', 'woocommerce' ),
             'type' => 'text',
             'description' => __( 'This controls the title which the user sees during checkout.', 'woocommerce' ),
-            'default' => __( 'Credit Card installment', 'woocommerce' ),
+            'default' => __( 'Credit Card Discount!', 'woocommerce' ),
             'desc_tip'      => true,
           ),
           'description' => array(
             'title' => __( 'Customer Message', 'woocommerce' ),
             'type' => 'textarea',
             'description' => __( 'This controls the description which the user sees during checkout', 'woocommerce' ),
-            'default' => 'Minimal transaction amount allowed for credit card installment is Rp. '.$this->get_option( 'min_amount' ). '</br> You will be redirected to fullpayment page if the transaction amount below this value'
+            'default' => 'Discount will be applied on the next step :)'
           ),
           'select_veritrans_environment' => array(
             'title' => __( 'Environment', 'woocommerce' ),
@@ -137,34 +138,41 @@
             'default' => '',
             'class' => 'production_settings sensitive'
           ),
-          'enable_bni_installment' => array(
-            'title' => __( 'Enable BNI installment ', 'woocommerce' ),
-            'type' => 'checkbox',
-            'label' => __( 'Enable BNI installment?', 'woocommerce' ),
-            'description' => __( 'Please contact us if you wish to enable this feature in the Production environment.', 'woocommerce' ),
-            'default' => 'no'
-          ),
-          'enable_bni_installment_terms' => array(
-            'title' => __( 'BNI Installment terms', 'woocommerce' ),
-            'type' => 'text',
-            'label' => __( 'BNI Installment terms', 'woocommerce' ),
-            'description' => __( 'use comma to separate value. Only 6 and 12 months term available on sandbox environment.', 'woocommerce' ),
-            'default' => '6,12'
-          ),
-          'enable_mandiri_installment' => array(
-            'title' => __( 'Enable Mandiri installment ', 'woocommerce' ),
-            'type' => 'checkbox',
-            'label' => __( 'Enable Mandiri installment?', 'woocommerce' ),
-            'description' => __( 'Please contact us if you wish to enable this feature in the Production environment.', 'woocommerce' ),
-            'default' => 'no'
-          ),
-          'enable_mandiri_installment_terms' => array(
-            'title' => __( 'Mandiri Installment terms', 'woocommerce'),
-            'type' => 'text',
-            'label' => __( 'Mandiri Installment terms', 'woocommerce' ),
-            'description' => __( 'use comma to separate value. Only 6 and 12 months term available on sandbox environment.', 'woocommerce' ),
-            'default' => '6,12'
-          ),
+          // 'enable_bni_installment' => array(
+          //   'title' => __( 'Enable BNI installment ', 'woocommerce' ),
+          //   'type' => 'checkbox',
+          //   'label' => __( 'Enable BNI installment?', 'woocommerce' ),
+          //   'description' => __( 'Please contact us if you wish to enable this feature in the Production environment.', 'woocommerce' ),
+          //   'default' => 'no'
+          // ),
+          // 'enable_bni_installment_terms' => array(
+          //   'title' => __( 'BNI Installment terms', 'woocommerce' ),
+          //   'type' => 'text',
+          //   'label' => __( 'BNI Installment terms', 'woocommerce' ),
+          //   'description' => __( 'use comma to separate value. Only 6 and 12 months term available on sandbox environment.', 'woocommerce' ),
+          //   'default' => '6,12'
+          // ),
+          // 'enable_mandiri_installment' => array(
+          //   'title' => __( 'Enable Mandiri installment ', 'woocommerce' ),
+          //   'type' => 'checkbox',
+          //   'label' => __( 'Enable Mandiri installment?', 'woocommerce' ),
+          //   'description' => __( 'Please contact us if you wish to enable this feature in the Production environment.', 'woocommerce' ),
+          //   'default' => 'no'
+          // ),
+          // 'enable_mandiri_installment_terms' => array(
+          //   'title' => __( 'Mandiri Installment terms', 'woocommerce'),
+          //   'type' => 'text',
+          //   'label' => __( 'Mandiri Installment terms', 'woocommerce' ),
+          //   'description' => __( 'use comma to separate value. Only 6 and 12 months term available on sandbox environment.', 'woocommerce' ),
+          //   'default' => '6,12'
+          // // ),
+          // 'installment_terms' => array(
+          //   'title' => __( 'Offline Installment terms', 'woocommerce'),
+          //   'type' => 'text',
+          //   'label' => __( 'Offline Installment terms', 'woocommerce' ),
+          //   'description' => __( 'use comma to separate value. Only 6 and 12 months term available on sandbox environment.', 'woocommerce' ),
+          //   'default' => '6,12'
+          // ),
           'enable_bin_filter' => array(
             'title' => __( 'Bin Number', 'woocommerce'),
             'type' => 'text',
@@ -172,13 +180,13 @@
             'description' => __( 'Please contact us if you wish to enable this feature in the Production environment.', 'woocommerce' ),
             'default' => '4,5'
           ),
-          'min_amount' => array(
-            'title' => __( 'Minimal Transaction Amount', 'woocommerce'),
-            'type' => 'int',
-            'label' => __( 'Minimal Transaction Amount', 'woocommerce' ),
-            'description' => __( 'Minimal transaction amount allowed to be paid with installment. (amount in IDR, without comma or period) example: 500000 </br> if the transaction amount is below this value, customer will be redirected to Credit Card fullpayment page', 'woocommerce' ),
-            'default' => '500000'
-          ),
+          // 'min_amount' => array(
+          //   'title' => __( 'Minimal Transaction Amount', 'woocommerce'),
+          //   'type' => 'int',
+          //   'label' => __( 'Minimal Transaction Amount', 'woocommerce' ),
+          //   'description' => __( 'Minimal transaction amount allowed to be paid with installment. (amount in IDR, without comma or period) example: 500000 </br> if the transaction amount is below this value, customer will be redirected to Credit Card fullpayment page', 'woocommerce' ),
+          //   'default' => '500000'
+          // ),
           'enable_sanitization' => array(
             'title' => __( 'Enable Sanitization', 'woocommerce' ),
             'type' => 'checkbox',
@@ -224,7 +232,20 @@
         $cart = $woocommerce->cart;
 
         $order = new WC_Order( $order_id );     
-      
+        
+        // add discount
+        // WC()->cart->add_discount( 'veritrans' );
+        $cart->add_discount('veritrans');
+        $order->add_coupon( 'veritrans', WC()->cart->get_coupon_discount_amount( 'veritrans' ), WC()->cart->get_coupon_discount_tax_amount( 'veritrans' ) );
+        $order->set_total( WC()->cart->shipping_total, 'shipping' );
+        $order->set_total( WC()->cart->get_cart_discount_total(), 'cart_discount' );
+        $order->set_total( WC()->cart->get_cart_discount_tax_total(), 'cart_discount_tax' );
+        $order->set_total( WC()->cart->tax_total, 'tax' );
+        $order->set_total( WC()->cart->shipping_tax_total, 'shipping_tax' );
+        $order->set_total( WC()->cart->total );
+        // $order->add_coupon('veritrans',10000);
+        // end of add discount
+
         Veritrans_Config::$isProduction = ($this->environment == 'production') ? true : false;
         Veritrans_Config::$serverKey = (Veritrans_Config::$isProduction) ? $this->server_key_v2_production : $this->server_key_v2_sandbox;     
         Veritrans_Config::$is3ds = true;
@@ -239,9 +260,9 @@
         );
 
         $enabled_payments = array();
-        if ($this->enable_credit_card == 'yes'){
-          $enabled_payments[] = 'credit_card';
-        }
+        // if ($this->enable_credit_card == 'yes'){
+        //   $enabled_payments[] = 'credit_card';
+        // }
 
         $params['vtweb']['enabled_payments'] = 'credit_card';
         $bins = $this->bin_filter;
@@ -324,10 +345,10 @@
         }
 
         // Discount
-        if ( $cart->get_cart_discount_total() > 0) {
+        if ( $order->get_cart_discount() > 0) {
           $items[] = array(
             'id' => 'totaldiscount',
-            'price' => $cart->get_cart_discount_total() * -1,
+            'price' => $order->get_cart_discount() * -1,
             'quantity' => 1,
             'name' => 'Total Discount'
           );
@@ -348,38 +369,53 @@
           }
         }
 
-        $params['transaction_details']['gross_amount'] = $order->get_total();
-        error_log('bni'.$this->enable_bni);
-        error_log('mandiri'.$this->enable_mandiri);
-        if($this->enable_bni == 'yes' || $this->enable_mandiri == 'yes')
+        //calculate gross amount
+        $total_amount=0;
+        // error_log('print r items[]' . print_r($items,true)); //debugan
+        foreach ($items as $item) {
+          $total_amount+=($item['price']*$item['quantity']);
+          // error_log('|||| Per item[]' . print_r($item,true)); //debugan
+
+        }
+
+        $params['transaction_details']['gross_amount'] = $total_amount;
+        // error_log('bni'.$this->enable_bni);
+        // error_log('mandiri'.$this->enable_mandiri);
+        // if($this->enable_bni == 'yes' || $this->enable_mandiri == 'yes')
+        if(false)
         {  
+            $installment_terms = array();
+
             $payment_options = array(
             'installment' => array(
-              'required' => true
+              'required' => true,
+              'installment_terms' => new stdClass(),
+              'offline_installment_terms' => []
               )
             );
 
-            $installment_terms = array();
-
-            $term_bni = $this->bni_terms;
-            error_log('term bni '.$term_bni);
-            $term_bni_array = explode(',' , $term_bni);
+            // $term_bni = $this->bni_terms;
+            // error_log('term bni '.$term_bni);
+            // $term_bni_array = explode(',' , $term_bni);
             
-            if($term_bni == "yes" || $term_bni_array != null)
-            {
-              $installment_terms['bni'] = $term_bni_array;
-            }
+            // if($term_bni == "yes" || $term_bni_array != null)
+            // {
+            //   $installment_terms['bni'] = $term_bni_array;
+            // }
             
-            $term_mandiri =  $this->mandiri_terms;
-            error_log('term mandiri '.$term_mandiri);
-            $term_mandiri_array = explode(',' , $term_mandiri);
+            // $term_mandiri =  $this->mandiri_terms;
+            // error_log('term mandiri '.$term_mandiri);
+            // $term_mandiri_array = explode(',' , $term_mandiri);
 
-            if($term_mandiri == "yes" || $term_mandiri_array != null)
-            {
-              $installment_terms['mandiri'] = $term_mandiri_array;
-            }
- 
-           
+            // if($term_mandiri == "yes" || $term_mandiri_array != null)
+            // {
+            //   $installment_terms['mandiri'] = $term_mandiri_array;
+            // }
+            
+            $term =  $this->installment_terms;
+            error_log('============installment_terms '.$term);
+            $term_array = explode(',' , $term);
+            
         }        
 
         // sift through the entire item to ensure that currency conversion is applied
@@ -396,15 +432,15 @@
 
         $params['item_details'] = $items;
         
-        if($params['transaction_details']['gross_amount'] >= $this->min_amount)
-        {
-          $payment_options['installment']['installment_terms'] = $installment_terms;
-          $params['vtweb']['payment_options'] = $payment_options;
-        }
+        // if($params['transaction_details']['gross_amount'] >= $this->min_amount)
+        // {
+        //   $payment_options['installment']['offline_installment_terms'] = $term_array;
+        //   $params['vtweb']['payment_options'] = $payment_options;
+        // }
 
         $woocommerce->cart->empty_cart();
         error_log(print_r($params,TRUE));
-        error_log(json_encode($params));
+        // error_log(json_encode($params));
         return Veritrans_VtWeb::getRedirectionUrl($params);
       }